requirements.txt
openai
anthropic
groq
transformers
torch
vllm
python-dotenv
streamlit
pinecone-client